% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_ptFCReg.R
\name{fitted_ptFCReg}
\alias{fitted_ptFCReg}
\title{Fitted functional responses from a \code{ptFCReg} object.}
\usage{
fitted_ptFCReg(object)
}
\arguments{
\item{object}{An object of class \code{ptFCReg}, returned by \code{\link{ptFCReg}} or \code{\link{smPtFCRegCoef}}.}
}
\value{
An \eqn{n}-by-\eqn{m} matrix, where each row corresponds to one subject,
and each column corresponds to a time point in \code{object$tGrid}.
}
\description{
Fitted functional responses from a \code{ptFCReg} object.
}
\examples{
set.seed(1)
n <- 50
nGridIn <- 101
tGrid <- seq(0, 1, length.out=nGridIn) # Functional data support
muX1 <- tGrid * 2 # mean function for X_1
sigma <- 1
beta0 <- 0
beta <- rbind(cos(tGrid), 1.5 + sin(tGrid))
Z <- MASS::mvrnorm(n, rep(0, 2), diag(2))
X_1 <- Z[, 1, drop=FALSE] \%*\% matrix(1, 1, nGridIn) + matrix(muX1, n, nGridIn, byrow=TRUE)
epsilon <- rnorm(n, sd=sigma)
Y <- t(sapply(seq_len(n), function(i) {
  beta0 + beta[1,] * X_1[i, ] + beta[2,] * Z[i, 2] + epsilon[i]
}))
dat <- list(X1=X_1, Z1=Z[, 2], Y=Y)
res <- ptFCReg(tGrid = tGrid, dat = dat)
smres <- smPtFCRegCoef(res, bw = 2.5 / (nGridIn-1), kernel_type = 'epan')
fit_res <- stats::fitted(res)
fit_smres <- stats::fitted(smres)
}
